Edit delete authorization done. Also, factories and seeder made for Answer model class.

+++ app/Http/Controllers/QuestionsController.php
@@ -89,7 +89,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));
+    	if ($this->authorize('update', $question)) {
+		    return view('questions.edit', compact('question'));
+	    }
     }
 
     /**
@@ -101,12 +103,15 @@ class QuestionsController extends Controller
      */
     public function update(Request $request, Question $question)
     {
-        $question->update([
-        	'title'     =>      $request->title,
-	        'body'      =>      $request->body
-        ]);
-	    session()->flash("success", "Question has been modified successfully!");
-	    return redirect(route('questions.index'));
+    	if ($this->authorize('update', $question)) {
+		    $question->update([
+			    'title'     =>      $request->title,
+			    'body'      =>      $request->body
+		    ]);
+		    session()->flash("success", "Question has been modified successfully!");
+		    return redirect(route('questions.index'));
+	    }
+    	abort(403);
     }
 
     /**
@@ -117,8 +122,11 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
-    	$question->delete();
-	    session()->flash("success", "Question has been deleted successfully!");
-	    return redirect(route('questions.index'));
+    	if ($this->authorize('delete', $question)) {
+		    $question->delete();
+		    session()->flash("success", "Question has been deleted successfully!");
+		    return redirect(route('questions.index'));
+	    }
+    	abort(403);
     }
 }
